+++ src/test/java/kitchenpos/order/application/listner/TableGroupOrderValidatorTest.java
@@ -9,9 +9,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import kitchenpos.order.application.OrderRepository;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.exception.OrderException;
+import kitchenpos.order.domain.listner.TableGroupOrderValidator;
 import kitchenpos.support.ServiceIntegrationTest;
 import kitchenpos.table.application.dto.OrderTableDto;
 import kitchenpos.table.domain.OrderTable;
@@ -28,8 +27,6 @@ class TableGroupOrderValidatorTest extends ServiceIntegrationTest {
     private TableGroupOrderValidator eventListener;
     @Autowired
     private OrderTableRepository orderTableRepository;
-    @Autowired
-    private OrderRepository orderRepository;
 
     @Test
     @DisplayName("Table의 주문 상태가 완료상태가 아닌경우 예외처리")
@@ -41,7 +38,8 @@ class TableGroupOrderValidatorTest extends ServiceIntegrationTest {
         );
 
         //when
-        assertThatThrownBy(() -> eventListener.validateUngroup(tableGroupDto.getId()))
+        final Long tableGroupId = tableGroupDto.getId();
+        assertThatThrownBy(() -> eventListener.validateUngroup(tableGroupId))
             .isInstanceOf(OrderException.class)
             .hasMessage(ORDER_IS_NOT_COMPLETION.getMessage());
     }
