+++ src/test/java/kitchenpos/table/application/TableGroupEventListenerTest.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Set;
 import kitchenpos.support.ServiceIntegrationTest;
-import kitchenpos.table.application.listener.TableGroupEventListener;
+import kitchenpos.table.domain.listener.TableGroupEventListener;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.table_group.application.TableGroupCreateEvent;
